Simplify |PDFLinkService.getDestinationHash| to return a stringified version of the destination array for explicit destinations
Currently for explicit destinations, compared to named destinations, we manually try to build a hash that often times is a quite poor representation of the *actual* destination. For PDF files using explicit destinations, this can make it difficult to accurately link to a specific section of the document through a URL.

With the added, and already existing, type checks in place for destinations, I don't really think that this patch exposes any "sensitive" internal destination code not already accessible through normal hash parameters.

*Please note:* just trying to improve the algorithm that generates the hash is unfortunately not possible in general, since there are cases where it will simply never work correctly (e.g. explicit `FitR` destinations).